Add symbolic link SAGE_DOC/output -> SAGE_DOC_OUTPUT

+++ src/sage_setup/docbuild/__main__.py
@@ -1,10 +1,13 @@
 from . import main
 main()
 
-# Remove old documentation output
+# Replace old documentation output by symbolic link to SAGE_DOC_OUTPUT
 import os
-from sage.env import SAGE_DOC
+from sage.env import SAGE_DOC, SAGE_DOC_OUTPUT
 old_doc = os.path.join(SAGE_DOC, "output")
-if os.path.exists(old_doc):
-    from shutil import rmtree
-    rmtree(old_doc)
+
+if not os.path.islink(old_doc):
+    if os.path.exists(old_doc):
+        from shutil import rmtree
+        rmtree(old_doc)
+    os.symlink(SAGE_DOC_OUTPUT, old_doc)
